requirements.txt
fastapi
uvicorn[standard]
twilio
openai
python-dotenv
groq